feat: Add trending sports event tickets page with filtering and statistics
- Implemented a new Blade view for trending sports event tickets.
- Added statistics overview including total trending tickets, active platforms, and average price.
- Included filters for sport type and results limit.
- Created a responsive grid layout for displaying trending tickets with detailed information.
- Added quick action buttons for user engagement.
- Implemented auto-refresh functionality for trending data.
- Enhanced navigation header with user menu and search functionality.

refactor: Update header and navigation layout components

- Refactored Header component to include new navigation items and user menu.
- Improved mobile navigation handling and search functionality.
- Added notification count display in the header.
- Updated NavigationLayout to manage user authentication and mobile menu state.

test: Create comprehensive test for ticket details system

- Developed a test script to validate ticket details functionality.
- Verified API endpoint responses and data structure completeness.
- Checked for the existence of necessary CSS and Blade view files.
- Ensured that all required fields are present in API responses.

+++ app/Models/ScrapedTicket.php
@@ -37,6 +37,7 @@ use Illuminate\Support\Str;
  * @property Carbon        $created_at
  * @property Carbon        $updated_at
  * @property Category|null $category
+ * @property float         $price
  * @property string        $formatted_price
  * @property float         $total_price
  * @property bool          $is_recent
@@ -241,9 +242,9 @@ class ScrapedTicket extends Model
     /**
      * Get price attribute (defaults to min_price for compatibility)
      */
-    public function getPriceAttribute(): int
+    public function getPriceAttribute(): float
     {
-        return $this->min_price ?? 0;
+        return (float) ($this->min_price ?? 0);
     }
 
     /**
@@ -259,9 +260,9 @@ class ScrapedTicket extends Model
     /**
      * Get  total price attribute
      */
-    public function getTotalPriceAttribute(): int
+    public function getTotalPriceAttribute(): float
     {
-        return $this->max_price ?? $this->min_price ?? 0;
+        return (float) ($this->max_price ?? $this->min_price ?? 0);
     }
 
     public function getIsRecentAttribute()
